requirements/airflow.txt
PyYAML>=6.0.2,<7
requests>=2.32.3,<3
pandas>=2.1.2,<2.2
pyarrow==16.1.0
sqlalchemy>=1.4.36,<2.0
psycopg2-binary>=2.9.9,<3
pyspark==3.5.2
